+++ src/Link.cc
@@ -17,6 +17,7 @@
 
 #include <ignition/msgs/Utility.hh>
 
+#include "ignition/gazebo/components/AngularAcceleration.hh"
 #include "ignition/gazebo/components/AngularVelocity.hh"
 #include "ignition/gazebo/components/CanonicalLink.hh"
 #include "ignition/gazebo/components/Collision.hh"
@@ -246,6 +247,14 @@ void Link::EnableVelocityChecks(EntityComponentManager &_ecm, bool _enable)
 }
 
 //////////////////////////////////////////////////
+std::optional<math::Vector3d> Link::WorldAngularAcceleration(
+    const EntityComponentManager &_ecm) const
+{
+  return _ecm.ComponentData<components::WorldAngularAcceleration>(
+      this->dataPtr->id);
+}
+
+//////////////////////////////////////////////////
 std::optional<math::Vector3d> Link::WorldLinearAcceleration(
     const EntityComponentManager &_ecm) const
 {
@@ -257,6 +266,10 @@ std::optional<math::Vector3d> Link::WorldLinearAcceleration(
 void Link::EnableAccelerationChecks(EntityComponentManager &_ecm, bool _enable)
     const
 {
+  enableComponent<components::WorldAngularAcceleration>(_ecm, this->dataPtr->id,
+      _enable);
+  enableComponent<components::AngularAcceleration>(_ecm, this->dataPtr->id,
+      _enable);
   enableComponent<components::WorldLinearAcceleration>(_ecm, this->dataPtr->id,
       _enable);
   enableComponent<components::LinearAcceleration>(_ecm, this->dataPtr->id,
@@ -320,6 +333,29 @@ void Link::AddWorldForce(EntityComponentManager &_ecm,
   // compute the resulting force and torque on the link origin.
   auto posComWorldCoord =
       worldPose->Data().Rot().RotateVector(inertial->Data().Pose().Pos());
+
+  math::Vector3d torque = posComWorldCoord.Cross(_force);
+
+  this->AddWorldWrench(_ecm, _force, torque);
+}
+
+//////////////////////////////////////////////////
+void Link::AddWorldForce(EntityComponentManager &_ecm,
+                         const math::Vector3d &_force,
+                         const math::Vector3d &_position) const
+{
+  auto inertial = _ecm.Component<components::Inertial>(this->dataPtr->id);
+  auto worldPose = _ecm.Component<components::WorldPose>(this->dataPtr->id);
+
+  // Can't apply force if the inertial's pose is not found
+  if (!inertial || !worldPose)
+    return;
+
+  // We want the force to be applied at an offset from the center of mass, but
+  // ExternalWorldWrenchCmd applies the force at the link origin so we need to
+  // compute the resulting force and torque on the link origin.
+  auto posComWorldCoord = worldPose->Data().Rot().RotateVector(
+    _position + inertial->Data().Pose().Pos());
 
   math::Vector3d torque = posComWorldCoord.Cross(_force);