+++ test/integration/battery_plugin.cc
@@ -24,6 +24,7 @@
 #include <ignition/common/Util.hh>
 #include <ignition/common/Filesystem.hh>
 #include <ignition/transport/Node.hh>
+#include <ignition/utilities/ExtraTestMacros.hh>
 
 #include <sdf/Root.hh>
 #include <sdf/World.hh>
@@ -71,8 +72,9 @@ class BatteryPluginTest : public InternalFixture<::testing::Test>
 
 
 /////////////////////////////////////////////////
-// Single model consuming single battery
-TEST_F(BatteryPluginTest, SingleBattery)
+// Single model consuming single batter
+// See https://github.com/ignitionrobotics/ign-gazebo/issues/1175
+TEST_F(BatteryPluginTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(SingleBattery))
 {
   const auto sdfPath = common::joinPaths(std::string(PROJECT_SOURCE_PATH),
     "test", "worlds", "battery.sdf");
@@ -161,7 +163,8 @@ TEST_F(BatteryPluginTest, SingleBattery)
 
 /////////////////////////////////////////////////
 // Battery with  power draining topics
-TEST_F(BatteryPluginTest, PowerDrainTopic)
+// See https://github.com/ignitionrobotics/ign-gazebo/issues/1175
+TEST_F(BatteryPluginTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(PowerDrainTopic))
 {
   const auto sdfPath = common::joinPaths(std::string(PROJECT_SOURCE_PATH),
     "test", "worlds", "battery.sdf");