+++ test/integration/level_manager.cc
@@ -23,6 +23,7 @@
 
 #include <ignition/common/Console.hh>
 #include <ignition/common/Util.hh>
+#include <ignition/utilities/ExtraTestMacros.hh>
 #include <sdf/Box.hh>
 #include <sdf/Cylinder.hh>
 #include <sdf/Joint.hh>
@@ -178,7 +179,8 @@ class LevelManagerFixture : public InternalFixture<::testing::Test>
 
 /////////////////////////////////////////////////
 /// Check default level includes entities not included by other levels
-TEST_F(LevelManagerFixture, DefaultLevel)
+// See https://github.com/ignitionrobotics/ign-gazebo/issues/1175
+TEST_F(LevelManagerFixture, IGN_UTILS_TEST_DISABLED_ON_WIN32(DefaultLevel))
 {
   std::vector<std::set<std::string>> levelEntityNamesList;
 
@@ -224,7 +226,7 @@ TEST_F(LevelManagerFixture, DefaultLevel)
 ///////////////////////////////////////////////
 /// Check a level is loaded when a performer is inside a level
 /// Check a level is unloaded when a performer is outside a level
-TEST_F(LevelManagerFixture, LevelLoadUnload)
+TEST_F(LevelManagerFixture, IGN_UTILS_TEST_DISABLED_ON_WIN32(LevelLoadUnload))
 {
   ModelMover perf1(*this->server->EntityByName("sphere"));
   this->server->AddSystem(perf1.systemPtr);
@@ -274,7 +276,7 @@ TEST_F(LevelManagerFixture, LevelLoadUnload)
 
 ///////////////////////////////////////////////
 /// Check behaviour of level buffers
-TEST_F(LevelManagerFixture, LevelBuffers)
+TEST_F(LevelManagerFixture, IGN_UTILS_TEST_DISABLED_ON_WIN32(LevelBuffers))
 {
   ModelMover perf1(*this->server->EntityByName("sphere"));
   this->server->AddSystem(perf1.systemPtr);
@@ -312,7 +314,8 @@ TEST_F(LevelManagerFixture, LevelBuffers)
 
 ///////////////////////////////////////////////
 /// Check that multiple performers can load/unload multiple levels independently
-TEST_F(LevelManagerFixture, LevelsWithMultiplePerformers)
+TEST_F(LevelManagerFixture,
+       IGN_UTILS_TEST_DISABLED_ON_WIN32(LevelsWithMultiplePerformers))
 {
   ModelMover perf1(*this->server->EntityByName("sphere"));
   ModelMover perf2(*this->server->EntityByName("box"));
@@ -429,7 +432,8 @@ TEST_F(LevelManagerFixture, LevelsWithMultiplePerformers)
 
 ///////////////////////////////////////////////
 /// Check that buffers work properly with multiple performers
-TEST_F(LevelManagerFixture, LevelBuffersWithMultiplePerformers)
+TEST_F(LevelManagerFixture,
+       IGN_UTILS_TEST_DISABLED_ON_WIN32(LevelBuffersWithMultiplePerformers))
 {
   ModelMover perf1(*this->server->EntityByName("sphere"));
   ModelMover perf2(*this->server->EntityByName("box"));