+++ test/integration/user_commands.cc
@@ -50,7 +50,8 @@ class UserCommandsTest : public InternalFixture<::testing::Test>
 };
 
 /////////////////////////////////////////////////
-TEST_F(UserCommandsTest, Create)
+// See https://github.com/ignitionrobotics/ign-gazebo/issues/1175
+TEST_F(UserCommandsTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(Create))
 {
   // Start server
   ServerConfig serverConfig;
@@ -329,7 +330,7 @@ TEST_F(UserCommandsTest, Create)
 }
 
 /////////////////////////////////////////////////
-TEST_F(UserCommandsTest, Remove)
+TEST_F(UserCommandsTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(Remove))
 {
   // Start server
   ServerConfig serverConfig;
@@ -517,7 +518,7 @@ TEST_F(UserCommandsTest, Remove)
 }
 
 /////////////////////////////////////////////////
-TEST_F(UserCommandsTest, Pose)
+TEST_F(UserCommandsTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(Pose))
 {
   // Start server
   ServerConfig serverConfig;
@@ -691,7 +692,7 @@ TEST_F(UserCommandsTest, Pose)
 
 /////////////////////////////////////////////////
 // https://github.com/ignitionrobotics/ign-gazebo/issues/634
-TEST_F(UserCommandsTest, IGN_UTILS_TEST_DISABLED_ON_MAC(Light))
+TEST_F(UserCommandsTest, IGN_UTILS_TEST_ENABLED_ONLY_ON_LINUX(Light))
 {
   // Start server
   ServerConfig serverConfig;
@@ -921,10 +922,29 @@ TEST_F(UserCommandsTest, IGN_UTILS_TEST_DISABLED_ON_MAC(Light))
   EXPECT_NEAR(1.5, spotLightComp->Data().SpotInnerAngle().Radian(), 0.1);
   EXPECT_NEAR(0.3, spotLightComp->Data().SpotOuterAngle().Radian(), 0.1);
   EXPECT_NEAR(0.9, spotLightComp->Data().SpotFalloff(), 0.1);
+
+  // Test light_config topic
+  const std::string lightTopic = "/world/lights_command/light_config";
+
+  msgs::Light lightMsg;
+  lightMsg.set_name("spot");
+  ignition::msgs::Set(lightMsg.mutable_diffuse(),
+    ignition::math::Color(1.0f, 1.0f, 1.0f, 1.0f));
+
+  // Publish light config
+  auto pub = node.Advertise<msgs::Light>(lightTopic);
+  pub.Publish(lightMsg);
+
+  server.Run(true, 100, false);
+  // Sleep for a small duration to allow Run thread to start
+  IGN_SLEEP_MS(10);
+
+  EXPECT_EQ(math::Color(1.0f, 1.0f, 1.0f, 1.0f),
+    spotLightComp->Data().Diffuse());
 }
 
 /////////////////////////////////////////////////
-TEST_F(UserCommandsTest, Physics)
+TEST_F(UserCommandsTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(Physics))
 {
   // Start server
   ServerConfig serverConfig;